# openapi.yaml
openapi: 3.1.0
info:
  title: UARE Custom Backend API
  version: 1.0.0
  description: |
    Production API for the UARE enterprise platform — auth, billing, orgs,
    CAD execution, workflows, workers, analytics, patents, and ops quality.
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://api.uare.app
    description: Production
security:
  - cookieAuth: []
  - devHeaderAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uare_session
    devHeaderAuth:
      type: apiKey
      in: header
      name: x-dev-user-id
    opsBearerToken:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 320
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                full_name:
                  type: string
                  maxLength: 200
      responses:
        '201':
          description: User registered and session created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      user:
                        type: object
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 320
                password:
                  type: string
                  maxLength: 128
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      user:
                        type: object
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
...
  /auth/me:
    get:
      tags: [auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      user:
                        type: object
        '401':
          description: Not authenticated

  # ── Billing ───────────────────────────────────────────────────────────────
...